+++ server/lib/globalData.py
@@ -60,7 +60,10 @@ class GlobalData:
 		self.version = 0.504
 
 		# revision of the used server
-		self.rev = 0
+		self.rev = 1
+
+		# Used database layout version.
+		self.dbVersion = 1
 
 		# name of this server
 		self.name = "AlertR Server"